requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
e2b>=1.4.0
sse-starlette>=2.2.0
pydantic>=2.10.0
python-dotenv>=1.0.0
docker>=7.0.0
pyyaml>=6.0.0
